add permissions to account grade status models
closes EVAL-3338
flag=custom_gradebook_statuses

Also adds authorization checks for all GraphQL endpoints for these models.

Test Plan:
- Verify that only users with root account :manage permission are able to
  fetch, create, or update Grade Statuses via GraphQL for that root
  account. Note that statuses can be fetched in two different ways — as a
  connection to an account

  {
    __typename
    account(id: "1") {
      customGradeStatusesConnection {
        nodes {
          _id
          color
        }
      }
    }
  }

  or directly as a node:

  {
    __typename
    node(id: "2") {
      ... on CustomGradeStatus {
        color
        _id
      }
    }
  }

- Verify that all submissions and scores belonging to a custom grade status
  have their custom_grade_status_id nullified when the custom grade status
  is destroyed.

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/323249
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Christopher Soto <christopher.soto@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>

+++ app/graphql/mutations/delete_custom_grade_status.rb
@@ -26,12 +26,15 @@ module Mutations
       raise GraphQL::ExecutionError, "custom gradebook statuses feature flag is disabled" unless Account.site_admin.feature_enabled?(:custom_gradebook_statuses)
 
       custom_grade_status = CustomGradeStatus.active.find(input[:id])
+
+      unless custom_grade_status.grants_right?(current_user, session, :delete)
+        raise GraphQL::ExecutionError, I18n.t("Insufficient permissions")
+      end
+
       context[:deleted_models] ||= {}
       context[:deleted_models][:custom_grade_status] = custom_grade_status
       custom_grade_status_id = custom_grade_status.id
       custom_grade_status.deleted_by = current_user
-      # TODO: add multi-tenant shard test
-      # TODO: add authorization check (which also ensures correct root account)
       custom_grade_status.destroy
       { custom_grade_status_id: }
     rescue ActiveRecord::RecordNotFound
